+++ crates/telio-model/src/lib.rs
@@ -1,9 +1,9 @@
 #![deny(missing_docs)]
 //! Crate containing models of various components
-pub mod api_config;
 pub mod config;
 pub mod constants;
 pub mod event;
+pub mod features;
 pub mod mesh;
 pub mod validation;
