+++ crates/telio-traversal/src/endpoint_providers/stun.rs
@@ -618,7 +618,7 @@ impl<Wg: WireGuard, E: Backoff> State<Wg, E> {
                     wg_port,
                     &sockets.ext_socket,
                     &self.pong_event,
-                    telio_model::api_config::EndpointProvider::Stun,
+                    telio_model::features::EndpointProvider::Stun,
                 )
                 .await
         } else {
@@ -1444,7 +1444,7 @@ mod tests {
                 .pong(
                     msg.get_wg_port(),
                     &addr.ip(),
-                    telio_model::api_config::EndpointProvider::Stun,
+                    telio_model::features::EndpointProvider::Stun,
                 )
                 .unwrap();
 
@@ -1510,7 +1510,7 @@ mod tests {
                 .pong(
                     msg.get_wg_port(),
                     &addr.ip(),
-                    telio_model::api_config::EndpointProvider::Stun,
+                    telio_model::features::EndpointProvider::Stun,
                 )
                 .unwrap();
